confirm and delay on confirmation

+++ static/js/clearStorage.js
@@ -5,6 +5,11 @@ export function removeActivity(index) {
 }
 
 export function clearLocalStorage() {
-  localStorage.clear();
-  alert("Local storage cleared!");
+  if (confirm("Are you sure you want to permanently clear all data?")) {
+    localStorage.clear();
+    // introduce a delay to make the change more aparent
+    setTimeout(() => {
+      alert("Local storage cleared!");
+    }, 100);
+  }
 }
